+++ src/main/java/com/trip/diary/controller/TripController.java
@@ -7,7 +7,6 @@ import com.trip.diary.service.TripService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.Slice;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
@@ -19,9 +18,7 @@ import java.util.List;
 @RequiredArgsConstructor
 @RequestMapping("/trips")
 public class TripController {
-
     private final TripService tripService;
-    private final MemberSearchService memberSearchService;
 
     @PostMapping
     private ResponseEntity<CreateTripDto> createTrip(@Valid @RequestBody CreateTripForm form,
@@ -34,47 +31,5 @@ public class TripController {
                                                @Valid @RequestBody UpdateTripForm form,
                                                @AuthenticationPrincipal MemberPrincipal principal) {
         return ResponseEntity.ok(tripService.updateTrip(tripId, form, principal.getMember()));
-    }
-
-    @PutMapping("/{tripId}/bookmarks")
-    private ResponseEntity<Void> bookmark(@PathVariable Long tripId,
-                                          @AuthenticationPrincipal MemberPrincipal principal) {
-        tripService.bookmark(tripId, principal.getMember());
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping("/bookmarks")
-    private ResponseEntity<Slice<TripDto>> readBookmarks(@RequestParam int page,
-                                                         @AuthenticationPrincipal MemberPrincipal principal) {
-        return ResponseEntity.ok(tripService.readBookmarks(page, principal.getMember()));
-    }
-
-    @PutMapping("/{tripId}/members/{memberId}")
-    private ResponseEntity<Void> inviteOrCancel(@PathVariable Long tripId,
-                                                @PathVariable Long memberId,
-                                                @AuthenticationPrincipal MemberPrincipal principal) {
-        tripService.invite(tripId, memberId, principal.getMember());
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping("/{tripId}/members/{memberId}")
-    private ResponseEntity<Void> kickOut(@PathVariable Long tripId,
-                                         @PathVariable Long memberId,
-                                         @AuthenticationPrincipal MemberPrincipal principal) {
-        tripService.kickOut(tripId, memberId, principal.getMember());
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping("/members/search")
-    private ResponseEntity<List<MemberDto>> searchAddableMembers(@RequestParam String keyword,
-                                                                 @AuthenticationPrincipal MemberPrincipal principal) {
-        return ResponseEntity.ok(memberSearchService.searchAddableMembers(keyword, principal.getMember()));
-    }
-
-    @GetMapping("/{tripId}/members/search")
-    private ResponseEntity<List<MemberDto>> searchAddableMembersInTrip(@PathVariable Long tripId,
-                                                                       @RequestParam String keyword,
-                                                                       @AuthenticationPrincipal MemberPrincipal principal) {
-        return ResponseEntity.ok(memberSearchService.searchAddableMembersInTrip(tripId, keyword, principal.getMember()));
     }
 }
