+++ src/test/java/com/trip/diary/service/ParticipantServiceTest.java
@@ -4,11 +4,13 @@ import com.trip.diary.domain.constants.ParticipantType;
 import com.trip.diary.domain.model.Member;
 import com.trip.diary.domain.model.Participant;
 import com.trip.diary.domain.model.Trip;
+import com.trip.diary.domain.repository.MemberRepository;
 import com.trip.diary.domain.repository.ParticipantRepository;
 import com.trip.diary.domain.repository.TripRepository;
 import com.trip.diary.dto.ParticipantDto;
 import com.trip.diary.dto.TripDto;
 import com.trip.diary.exception.ErrorCode;
+import com.trip.diary.exception.MemberException;
 import com.trip.diary.exception.ParticipantException;
 import com.trip.diary.exception.TripException;
 import org.junit.jupiter.api.DisplayName;
@@ -36,6 +38,11 @@ class ParticipantServiceTest {
 
     @Mock
     private ParticipantRepository participantRepository;
+    @Mock
+    private MemberRepository memberRepository;
+
+    @Mock
+    private NotificationService notificationService;
 
     @Mock
     private ApplicationEventPublisher applicationEventPublisher;
@@ -102,11 +109,11 @@ class ParticipantServiceTest {
         //when
         List<ParticipantDto> result = participantService.getTripParticipants(1L, member);
         //then
-        assertTrue(result.get(0).isReader());
-        assertTrue(result.get(0).isAccepted());
+        assertTrue(result.get(0).getIsReader());
+        assertTrue(result.get(0).getIsAccepted());
         assertEquals(participant1.getNickname(), result.get(1).getNickname());
-        assertFalse(result.get(1).isReader());
-        assertFalse(result.get(1).isAccepted());
+        assertFalse(result.get(1).getIsReader());
+        assertFalse(result.get(1).getIsAccepted());
     }
 
     @Test
@@ -138,11 +145,11 @@ class ParticipantServiceTest {
         //when
         List<ParticipantDto> result = participantService.getTripParticipants(1L, participant1);
         //then
-        assertFalse(result.get(0).isReader());
-        assertTrue(result.get(0).isAccepted());
+        assertFalse(result.get(0).getIsReader());
+        assertTrue(result.get(0).getIsAccepted());
         assertNotEquals(participant1.getNickname(), result.get(1).getNickname());
-        assertTrue(result.get(1).isReader());
-        assertFalse(result.get(1).isAccepted());
+        assertTrue(result.get(1).getIsReader());
+        assertFalse(result.get(1).getIsAccepted());
     }
 
     @Test
@@ -182,11 +189,11 @@ class ParticipantServiceTest {
         //when
         List<ParticipantDto> result = participantService.getTripParticipants(1L, member);
         //then
-        assertFalse(result.get(0).isReader());
-        assertTrue(result.get(0).isAccepted());
+        assertFalse(result.get(0).getIsReader());
+        assertTrue(result.get(0).getIsAccepted());
         assertEquals(participant1.getNickname(), result.get(1).getNickname());
-        assertFalse(result.get(1).isReader());
-        assertFalse(result.get(1).isAccepted());
+        assertFalse(result.get(1).getIsReader());
+        assertFalse(result.get(1).getIsAccepted());
     }
 
     @Test
@@ -264,8 +271,8 @@ class ParticipantServiceTest {
         List<TripDto> result = participantService.getInvitedTripList(member);
         //then
         assertEquals(1L, result.get(0).getId());
-        assertTrue(result.get(0).getParticipants().get(0).isAccepted());
-        assertFalse(result.get(0).getParticipants().get(1).isAccepted());
+        assertTrue(result.get(0).getParticipants().get(0).getIsAccepted());
+        assertFalse(result.get(0).getParticipants().get(1).getIsAccepted());
     }
 
     @Test
@@ -349,6 +356,306 @@ class ParticipantServiceTest {
                 () -> participantService.denyTripInvitation(1L, member));
         //then
         assertEquals(ErrorCode.NOT_INVITED_TRIP, exception.getErrorCode());
+    }
+
+    @Test
+    @DisplayName("여행 기록장에 멤버 초대 성공")
+    void invite_success() {
+        //given
+        Member target = Member.builder()
+                .id(4L)
+                .username("1234qwert")
+                .nickname("오땡땡")
+                .password("1234567")
+                .profilePath(null)
+                .phone("01011111114")
+                .build();
+
+        Trip trip = Trip.builder()
+                .id(1L)
+                .title("임의의 타이틀")
+                .isPrivate(true)
+                .description("임의의 설명")
+                .leader(member)
+                .participants(List.of(
+                        Participant.builder()
+                                .member(member)
+                                .type(ParticipantType.ACCEPTED)
+                                .build(),
+                        Participant.builder()
+                                .member(participant1)
+                                .type(ParticipantType.PENDING)
+                                .build(),
+                        Participant.builder()
+                                .member(participant2)
+                                .type(ParticipantType.PENDING)
+                                .build()))
+                .build();
+        given(tripRepository.findById(anyLong())).willReturn(Optional.of(trip));
+        given(memberRepository.findById(anyLong())).willReturn(Optional.of(target));
+        given(participantRepository.existsByTripAndMemberAndType(any(), any(), any())).willReturn(true);
+        given(participantRepository.existsByTripAndMember(any(), any())).willReturn(false);
+        //when
+        participantService.invite(1L, 4L, member);
+        //then
+        verify(participantRepository, times(1)).save(any());
+    }
+
+    @Test
+    @DisplayName("여행 기록장에 멤버 초대 실패 - 이미 초대된 유저임")
+    void invite_failWhenUserAlreadyInvited() {
+        //given
+        Member target = Member.builder()
+                .id(4L)
+                .username("1234qwert")
+                .nickname("오땡땡")
+                .password("1234567")
+                .profilePath(null)
+                .phone("01011111114")
+                .build();
+
+        Trip trip = Trip.builder()
+                .id(1L)
+                .title("임의의 타이틀")
+                .isPrivate(true)
+                .description("임의의 설명")
+                .leader(member)
+                .participants(List.of(
+                        Participant.builder()
+                                .member(member)
+                                .type(ParticipantType.ACCEPTED)
+                                .build(),
+                        Participant.builder()
+                                .member(participant1)
+                                .type(ParticipantType.PENDING)
+                                .build(),
+                        Participant.builder()
+                                .member(participant2)
+                                .type(ParticipantType.PENDING)
+                                .build()))
+                .build();
+        given(tripRepository.findById(anyLong())).willReturn(Optional.of(trip));
+        given(memberRepository.findById(anyLong())).willReturn(Optional.of(target));
+        given(participantRepository.existsByTripAndMemberAndType(any(), any(), any())).willReturn(true);
+        given(participantRepository.existsByTripAndMember(any(), any())).willReturn(true);
+        //when
+        TripException exception = assertThrows(TripException.class,
+                () -> participantService.invite(1L, 4L, member));
+        //then
+        assertEquals(exception.getErrorCode(), ErrorCode.USER_ALREADY_INVITED);
+    }
+
+    @Test
+    @DisplayName("여행 기록장에 멤버 초대 실패 - 해당 여행 기록장 없음")
+    void invite_failWhenNotFoundTrip() {
+        //given
+        given(tripRepository.findById(anyLong())).willReturn(Optional.empty());
+        //when
+        TripException exception = assertThrows(TripException.class,
+                () -> participantService.invite(1L, 4L, member));
+        //then
+        assertEquals(exception.getErrorCode(), ErrorCode.NOT_FOUND_TRIP);
+    }
+
+    @Test
+    @DisplayName("여행 기록장에 멤버 초대 실패 - 해당 유저 없음")
+    void invite_failWhenNotFoundMember() {
+        //given
+        Trip trip = Trip.builder()
+                .id(1L)
+                .title("임의의 타이틀")
+                .isPrivate(true)
+                .description("임의의 설명")
+                .leader(member)
+                .participants(List.of(
+                        Participant.builder()
+                                .member(member)
+                                .type(ParticipantType.ACCEPTED)
+                                .build(),
+                        Participant.builder()
+                                .member(participant1)
+                                .type(ParticipantType.PENDING)
+                                .build(),
+                        Participant.builder()
+                                .member(participant2)
+                                .type(ParticipantType.PENDING)
+                                .build()))
+                .build();
+        given(tripRepository.findById(anyLong())).willReturn(Optional.of(trip));
+        given(memberRepository.findById(anyLong())).willReturn(Optional.empty());
+        //when
+        MemberException exception = assertThrows(MemberException.class,
+                () -> participantService.invite(1L, 4L, member));
+        //then
+        assertEquals(exception.getErrorCode(), ErrorCode.NOT_FOUND_MEMBER);
+    }
+
+    @Test
+    @DisplayName("여행 기록장에 멤버 초대 실패 - 해당 권한 없음")
+    void invite_failWhenNotAuthorityWriteTrip() {
+        //given
+        Member target = Member.builder()
+                .id(4L)
+                .username("1234qwert")
+                .nickname("오땡땡")
+                .password("1234567")
+                .profilePath(null)
+                .phone("01011111114")
+                .build();
+
+        Trip trip = Trip.builder()
+                .id(1L)
+                .title("임의의 타이틀")
+                .isPrivate(true)
+                .description("임의의 설명")
+                .leader(participant1)
+                .participants(List.of(
+                        Participant.builder()
+                                .member(participant1)
+                                .type(ParticipantType.PENDING)
+                                .build(),
+                        Participant.builder()
+                                .member(participant2)
+                                .type(ParticipantType.PENDING)
+                                .build()))
+                .build();
+        given(tripRepository.findById(anyLong())).willReturn(Optional.of(trip));
+        given(memberRepository.findById(anyLong())).willReturn(Optional.of(target));
+        given(participantRepository.existsByTripAndMemberAndType(any(), any(), any())).willReturn(false);
+        //when
+        TripException exception = assertThrows(TripException.class,
+                () -> participantService.invite(1L, 4L, member));
+        //then
+        assertEquals(exception.getErrorCode(), ErrorCode.NOT_AUTHORITY_WRITE_TRIP);
+    }
+
+    @Test
+    @DisplayName("여행 기록장에 멤버 퇴출 성공")
+    void kickOut_success() {
+        //given
+        Member target = Member.builder()
+                .id(4L)
+                .username("1234qwert")
+                .nickname("오땡땡")
+                .password("1234567")
+                .profilePath(null)
+                .phone("01011111114")
+                .build();
+
+        Trip trip = Trip.builder()
+                .id(1L)
+                .title("임의의 타이틀")
+                .isPrivate(true)
+                .description("임의의 설명")
+                .leader(member)
+                .participants(List.of(
+                        Participant.builder()
+                                .member(member)
+                                .type(ParticipantType.ACCEPTED)
+                                .build(),
+                        Participant.builder()
+                                .member(participant1)
+                                .type(ParticipantType.PENDING)
+                                .build(),
+                        Participant.builder()
+                                .member(participant2)
+                                .type(ParticipantType.PENDING)
+                                .build()))
+                .build();
+        given(tripRepository.findById(anyLong())).willReturn(Optional.of(trip));
+        given(participantRepository.findByTripAndMember_Id(any(), anyLong()))
+                .willReturn(Optional.of(Participant.builder()
+                        .member(participant2)
+                        .type(ParticipantType.PENDING)
+                        .build()
+                ));
+        //when
+        participantService.kickOut(1L, 4L, member);
+        //then
+        verify(participantRepository, times(1)).delete(any());
+    }
+
+    @Test
+    @DisplayName("여행 기록장에 멤버 퇴출 실패 - 해당 여행 기록장 없음")
+    void kickOut_failWhenNotFoundTrip() {
+        //given
+        given(tripRepository.findById(anyLong())).willReturn(Optional.empty());
+        //when
+        TripException exception = assertThrows(TripException.class,
+                () -> participantService.invite(1L, 4L, member));
+        //then
+        assertEquals(exception.getErrorCode(), ErrorCode.NOT_FOUND_TRIP);
+    }
+
+    @Test
+    @DisplayName("여행 기록장에 멤버 퇴출 실패 - 해당 유저 없음")
+    void kickOut_failWhenNotFoundMember() {
+        //given
+        Trip trip = Trip.builder()
+                .id(1L)
+                .title("임의의 타이틀")
+                .isPrivate(true)
+                .description("임의의 설명")
+                .leader(member)
+                .participants(List.of(
+                        Participant.builder()
+                                .member(member)
+                                .type(ParticipantType.ACCEPTED)
+                                .build(),
+                        Participant.builder()
+                                .member(participant1)
+                                .type(ParticipantType.PENDING)
+                                .build(),
+                        Participant.builder()
+                                .member(participant2)
+                                .type(ParticipantType.PENDING)
+                                .build()))
+                .build();
+        given(tripRepository.findById(anyLong())).willReturn(Optional.of(trip));
+        given(participantRepository.findByTripAndMember_Id(any(), anyLong()))
+                .willReturn(Optional.empty());
+        //when
+        TripException exception = assertThrows(TripException.class,
+                () -> participantService.kickOut(1L, 4L, member));
+        //then
+        assertEquals(exception.getErrorCode(), ErrorCode.NOT_FOUND_PARTICIPANT);
+    }
+
+    @Test
+    @DisplayName("여행 기록장에 멤버 퇴출 실패 - 해당 권한 없음")
+    void kickOut_failWhenNotAuthorityWriteTrip() {
+        //given
+        Member target = Member.builder()
+                .id(4L)
+                .username("1234qwert")
+                .nickname("오땡땡")
+                .password("1234567")
+                .profilePath(null)
+                .phone("01011111114")
+                .build();
+
+        Trip trip = Trip.builder()
+                .id(1L)
+                .title("임의의 타이틀")
+                .isPrivate(true)
+                .description("임의의 설명")
+                .leader(participant1)
+                .participants(List.of(
+                        Participant.builder()
+                                .member(participant1)
+                                .type(ParticipantType.PENDING)
+                                .build(),
+                        Participant.builder()
+                                .member(participant2)
+                                .type(ParticipantType.PENDING)
+                                .build()))
+                .build();
+        given(tripRepository.findById(anyLong())).willReturn(Optional.of(trip));
+        //when
+        TripException exception = assertThrows(TripException.class,
+                () -> participantService.kickOut(1L, 4L, member));
+        //then
+        assertEquals(exception.getErrorCode(), ErrorCode.NOT_AUTHORITY_WRITE_TRIP);
     }
 
 }
